WIP options premium for ETH pair
