+++ contracts/tests/TestOptionsPremiumPricer.sol
@@ -47,16 +47,38 @@ contract TestOptionsPremiumPricer is OptionsPremiumPricer {
     function testGetPremiumInStables(
         uint256 st,
         uint256 expiryTimestamp,
-        bool isPut,
-        bool spotInNative
+        bool isPut
     ) external view returns (uint256 result, uint256 gas) {
         bytes memory data =
             abi.encodeWithSelector(
                 this.getPremiumInStables.selector,
                 st,
                 expiryTimestamp,
-                isPut,
-                spotInNative
+                isPut
+            );
+
+        uint256 startgas = gasleft();
+        (bool success, bytes memory returnData) =
+            address(this).staticcall(data);
+        gas = startgas - gasleft();
+
+        result = 0;
+        if (success) {
+            result = abi.decode(returnData, (uint256));
+        }
+    }
+
+    function testGetPremiumNativePairsInStables(
+        uint256 st,
+        uint256 expiryTimestamp,
+        bool isPut
+    ) external view returns (uint256 result, uint256 gas) {
+        bytes memory data =
+            abi.encodeWithSelector(
+                this.getPremiumNativePairsInStables.selector,
+                st,
+                expiryTimestamp,
+                isPut
             );
 
         uint256 startgas = gasleft();
